Week3 - Exercise 1
This program is about the length of a movie. The movies should be around 90 minutes
so that they can be broadcast on a TV channel along with commercial breaks. The
user should enter the length of the movie in hours, minutes and seconds.
a. If the movie is shorter than 30 minutes, the program should print that it is a
“short movie”,
b. For movies up to 90 minutes, it should print “the movie can be broadcast on
TV as it is”,
c. For movies longer than 90 minutes, and less than 120 minutes, it should print
“should be cut so that it can be broadcast on TV.”
d. For movies longer than 120 minutes, it should print “it is not suitable for TV.”